# docs/swagger.yaml
basePath: /
definitions:
...
  models.MergeStatus:
    enum:
    - ""
    - pending
    - merged
    - failed
    - conflict
    type: string
    x-enum-varnames:
    - MergeStatusNone
    - MergeStatusPending
    - MergeStatusMerged
    - MergeStatusFailed
    - MergeStatusConflict
...
  models.SwarmRole:
    enum:
    - ""
    - swarm_parent
    - planner
    - worker
    - reviewer
    - merger
    - integrator
    - merger
    type: string
    x-enum-varnames:
    - SwarmRoleNone
    - SwarmRoleParent
    - SwarmRolePlanner
    - SwarmRoleWorker
    - SwarmRoleReviewer
    - SwarmRoleMerger
    - SwarmRoleLegacyIntegrator
    - SwarmRoleIntegrator
  models.Task:
    properties:
      agent_definition_id:
        description: 'Optional: agent definition (system prompt, skills, MCP)'
        type: string
      agent_id:
        description: 'Optional: LLM config (model) to use; uses default if nil'
        type: string
      auto_merge:
        type: boolean
      base_branch:
        description: Git branch this task should base its worktree on (from parent
          lineage)
        type: string
      base_commit_sha:
        description: Git commit SHA to base worktree on (from parent lineage)
        type: string
      category:
        $ref: '#/definitions/models.TaskCategory'
      chain_config:
        description: JSON configuration for task chaining
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      created_via:
        description: 'Origin: "web", "telegram", etc.'
        type: string
      display_order:
        type: integer
      goal_met:
        description: 'Derived: task''s persisted goal is achieved'
        type: boolean
      has_goal:
        description: 'Derived: task has a non-cleared persisted goal'
        type: boolean
      id:
        type: string
      lineage_depth:
        description: 'Depth in chain: 0 = root, 1 = child, 2 = grandchild, etc.'
        type: integer
      merge_status:
        $ref: '#/definitions/models.MergeStatus'
      merge_target_branch:
        type: string
      parent_task_id:
        description: 'Optional: references parent task for chaining or swarm grouping'
        type: string
      priority:
        type: integer
      project_id:
        type: string
      prompt:
        type: string
      status:
        $ref: '#/definitions/models.TaskStatus'
      swarm_children:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      swarm_config:
        type: string
      swarm_role:
        $ref: '#/definitions/models.SwarmRole'
      swarm_sequence:
        type: integer
      swarm_status:
        type: string
      tag:
        $ref: '#/definitions/models.TaskTag'
      telegram_chat_id:
        description: Telegram chat ID for sending completion notifications
        type: integer
      title:
        type: string
      updated_at:
        type: string
      worktree_branch:
        type: string
      worktree_path:
        type: string
    type: object
  models.TaskCategory:
    enum:
    - active
    - completed
    - backlog
    - scheduled
    - chat
    type: string
    x-enum-comments:
      CategoryChat: Internal category for chat messages (not displayed in UI)
    x-enum-descriptions:
    - ""
    - ""
    - ""
    - ""
    - Internal category for chat messages (not displayed in UI)
    x-enum-varnames:
    - CategoryActive
    - CategoryCompleted
    - CategoryBacklog
    - CategoryScheduled
    - CategoryChat
  models.TaskStatus:
    enum:
    - pending
    - queued
    - running
    - completed
    - failed
    - cancelled
    - blocked
    type: string
    x-enum-comments:
      StatusBlocked: Waiting for parent task to complete (chained child pre-created
        for visibility)
    x-enum-descriptions:
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - Waiting for parent task to complete (chained child pre-created for visibility)
    x-enum-varnames:
    - StatusPending
    - StatusQueued
    - StatusRunning
    - StatusCompleted
    - StatusFailed
    - StatusCancelled
    - StatusBlocked
  models.TaskTag:
    enum:
    - ""
    - feature
    - bug
    type: string
    x-enum-varnames:
    - TagNone
    - TagFeature
    - TagBug
...
  update.ActiveWork:
    properties:
      automation_activities:
        type: integer
      chat_executions:
        type: integer
      task_executions:
        type: integer
    type: object
  update.CoordinatorSnapshot:
    properties:
      channel:
        type: string
      configuration_error:
        type: string
      current_version:
        type: string
      distribution:
        type: string
      drain:
        $ref: '#/definitions/update.DrainStatus'
      error:
        type: string
      manual:
        type: boolean
      release:
        $ref: '#/definitions/update.VerifiedRelease'
      staged:
        type: boolean
      state:
        type: string
    type: object
  update.DrainStatus:
    properties:
      active:
        $ref: '#/definitions/update.ActiveWork'
      expires_at:
        type: string
      generation:
        type: string
      queued_total:
        type: integer
      state:
        type: string
    type: object
  update.ReleaseMetadata:
    properties:
      channel:
        type: string
      commit:
        type: string
      database_compatibility:
        type: string
      expires_at:
        type: string
      minimum_hosted_version:
        type: string
      minimum_updater_version:
        type: string
      published_at:
        type: string
      release_notes_url:
        type: string
      schema_version:
        type: integer
      targets:
        items:
          $ref: '#/definitions/update.Target'
        type: array
      version:
        type: string
    type: object
  update.Target:
    properties:
      arch:
        type: string
      filename:
        type: string
      filetype:
        type: string
      id:
        type: string
      image_ref:
        type: string
      install_layout:
        type: string
      kind:
        type: string
      os:
        type: string
      purpose:
        type: string
      sha256:
        type: string
      size:
        type: integer
      url:
        type: string
      variant:
        type: string
    type: object
  update.VerifiedRelease:
    properties:
      action:
        type: string
      apply_supported:
        type: boolean
      metadata:
        $ref: '#/definitions/update.ReleaseMetadata'
      target:
        $ref: '#/definitions/update.Target'
    type: object
...
host: localhost:3001
info:
  contact:
    email: support@openvibely.dev
    name: API Support
    url: https://github.com/openvibely/openvibely
  description: |-
    REST API for OpenVibely - AI-powered task scheduling and management
    This API provides endpoints for managing projects, tasks, and chat interactions with AI agents.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: OpenVibely API
  version: "1.0"
paths:
...
  /api/system/update:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/update.CoordinatorSnapshot'
        "204":
          description: No Content
      summary: Get system update state
      tags:
      - system
  /api/system/update/apply:
    post:
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/update.CoordinatorSnapshot'
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      summary: Accept a system update
      tags:
      - system
  /api/system/update/cancel:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/update.CoordinatorSnapshot'
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      summary: Cancel a system update
      tags:
      - system
  /api/system/update/events:
    get:
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
          schema:
            type: string
        "204":
          description: No Content
      summary: Stream system update events
      tags:
      - system
...
  /api/tasks/{id}/swarm:
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
      summary: Get task swarm
      tags:
      - tasks
  /api/tasks/{id}/swarm/cancel:
    post:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cancel swarm
      tags:
      - tasks
  /api/tasks/{id}/swarm/followup:
    post:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Follow-up message
        in: formData
        name: message
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send swarm follow-up
      tags:
      - tasks
  /api/tasks/{id}/swarm/rerun-integrator:
    post:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Rerun swarm merger
      tags:
      - tasks
  /api/tasks/{id}/swarm/rerun-merger:
    post:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Rerun swarm merger
      tags:
      - tasks
  /api/tasks/{id}/swarm/rerun-reviewer:
    post:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Rerun swarm reviewer
      tags:
      - tasks
  /api/tasks/{id}/swarm/start:
    post:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start swarm planner
      tags:
      - tasks
schemes:
- http
- https
swagger: "2.0"
...